// src/types/project.ts
export interface Project {
  title: string;
  description: string;
  features: string[];
  image?: string;
  thumbnail?: string;
  liveLink?: string;
 githubServer?: string;
  githubClient?: string;
    technologies: string[];
    category?: string;
    date?: string;
    _id?: string;
    startDate?: string;
    endDate?: string;
    createdAt?: string;
    updatedAt?: string;
}